+++ src/main/webapp/@storybook/react.js
@@ -1,7 +1,7 @@
 import { storiesOf as of } from '@connexta/ace/@storybook/react'
 
 import withTheme from '../theme/with-theme'
-import withDrawProvider from '../location/with-draw-provider'
+import withDrawProvider from '../components/location/with-draw-provider'
 import withMuiPickersProvider from '../with-mui-pickers-provider'
 
 export const storiesOf = (name, m) => {
