+++ src/main/webapp/components/basic-search/basic-search-helper.js
@@ -1,5 +1,5 @@
 import { geoJSONToGeometryJSON } from 'geospatialdraw/bin/geometry/utilities'
-import { geoToFilter } from '../../location'
+import { geoToFilter } from '../location'
 import { makeSearchGeoIdForFilter } from '../query-builder/filter'
 const { Map, Set, fromJS } = require('immutable')
 export const APPLY_TO_KEY = 'applyTo'
