--- src/main/webapp/components/containers/index.js
@@ -1,3 +1,0 @@
-import FilterCard from './filter-card'
-
-export { FilterCard }
