+++ src/main/webapp/components/containers/index.tsx
@@ -1,0 +1,3 @@
+const FilterCard = require('./filter-card')
+import SpacedLinearContainer from './spaced-linear-container'
+export { FilterCard, SpacedLinearContainer }
