+++ src/main/webapp/components/location/coordinate-value.tsx
