+++ src/main/webapp/components/location/geo-editor.tsx
