+++ src/main/webapp/components/location/keyword/props.tsx
