+++ src/main/webapp/components/location/point/index.tsx
