+++ src/main/webapp/components/location/point/props.tsx
