+++ src/main/webapp/components/location/polygon.tsx
